Extract InMemoryHistoryTestBase
